Drop support for End-of-Lifed Rails versions
At the time of pull request submission, [official Rails support for
versions `4.2`, `5.0`, and `5.1` have ended][maintenance_policy]

This commit configures the continuous integration test suite to execute
against `5.2` and `6.0`.

To support those versions in our continuous integration containers,
remove configuration code to limit the `bundler` version to `< 2.0`.

Additionally, this commit amends the `README` to mention `5.2` as the
minimally supported version.

[maintenance_policy]: https://guides.rubyonrails.org/maintenance_policy.html